feat: update error processing in tests, feat: change int to size_t

+++ nurseryRhyme/list.h
@@ -1,6 +1,8 @@
 #pragma once
 
-typedef int Value;
+#include <stdlib.h>
+
+typedef size_t Value;
 
 typedef struct Node Node;
